refactoring the logic
