+++ app/client/src/ee/PluginActionEditor/hooks/useAnalyticsOnRunClick.ts
@@ -1,0 +1,1 @@
+export * from "ce/PluginActionEditor/hooks/useAnalyticsOnRunClick";
