--- app/client/src/ee/PluginActionEditor/utils/callRunActionAnalytics.ts
@@ -1,1 +1,0 @@
-export * from "ce/PluginActionEditor/utils/callRunActionAnalytics";
